ZEPPELIN-1595. Make ZeppelinContext extensible
